Extract workflow node execution helpers
